server/requirements.txt
asgiref==3.6.0
black==23.3.0
boto3==1.26.123
botocore==1.29.123
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-auth==2.17.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
